+++ Examples/PSPDFCatalog/EmbeddedExample/PSCAnnotationTestController.m
@@ -174,7 +174,7 @@
     return resolvedPath;
 }
 
-- (void)pdfViewController:(PSPDFViewController *)pdfController willShowController:(id)viewController embeddedInController:(id)controller animated:(BOOL)animated {
+- (BOOL)pdfViewController:(PSPDFViewController *)pdfController shouldShowController:(id)viewController embeddedInController:(id)controller animated:(BOOL)animated {
     NSLog(@"willShowViewController: %@ embeddedIn:%@ animated: %d", viewController, controller, animated);
     
     // example how to intercept PSPDFSearchViewController and change the barStyle to black
@@ -183,6 +183,7 @@
         searchController.searchBar.barStyle = UIBarStyleBlack;
         searchController.searchBar.tintColor = [UIColor blackColor];
     }
+    return YES;
 }
 
 - (void)pdfViewController:(PSPDFViewController *)pdfController didShowController:(id)viewController embeddedInController:(id)controller animated:(BOOL)animated {